refactor(watch): extract useDropProgressPoll from useAppModel
